phonesbook/requirements.txt
asgiref==3.3.4
csvImporter==0.1.3.2
Django==3.2
django-crispy-forms==1.11.2
django-phonenumber-field==5.1.0
loading==0.0.1
phonenumbers==8.12.22
phonenumberslite==8.12.22
pytz==2021.1
sqlparse==0.4.1
